requirements.txt
Django==5.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.7.0
dj-database-url==2.2.0
python-dotenv==1.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
